UltraBlame original commit: 40b7b08eaa123caee88b3e7c59b8c4e811a50898

+++ js/src/jsobj.h
@@ -621,7 +621,7 @@ struct JSObject : js::gc::Cell {
 
      unsigned finalizeKind() const;
 
-    uint32 numSlots() const { return capacity; }
+    uint32 numSlots() const { return uint32(capacity); }
 
     inline size_t structSize() const;
     inline size_t slotsAndStructSize() const;
